apps/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0
python-multipart>=0.0.6
slowapi>=0.1.5
pyotp>=2.8.0
qrcode>=7.4.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
